+++ co-cli/tests/cli/mod.rs
@@ -12,6 +12,7 @@ pub fn co_command() -> Command {
 pub mod basic;
 pub mod content;
 pub mod context;
+pub mod create;
 pub mod i18n;
 pub mod lead;
 pub mod locate;